Create GMM models for enroll and recognize enrolled user

+++ java/SpeakerRecognizer/src/main/java/speakerid/controller/HomeController.java
@@ -27,18 +27,23 @@ import java.util.*;
 public class HomeController {
     private Map<String, Map<Integer, String>> speakers = new HashMap<>();
     private Map<String, List<String>> texts = new HashMap<>();
+
     private static Socket clientSocket;
     private static BufferedReader input;
     private static DataOutputStream output;
 
     private static final String HOST = "logic";
     private static final int PORT = 1024;
+    private static final String USER_DICTOR_NUMBER = "30";
 
     @Autowired private ResourceLoader loader;
     @Autowired private Cleanup cleanup;
 
     @Value("${speakers.directory}")
     private String speakersDirectory;
+
+    @Value("${kaldi.directory}")
+    private String kaldiDirectory;
 
     @Value("${speakers.filename}")
     private String speakersFilename;
@@ -61,6 +66,8 @@ public class HomeController {
             }
             input = new BufferedReader(new InputStreamReader(clientSocket.getInputStream()));
             output = new DataOutputStream(clientSocket.getOutputStream());
+            new ProcessBuilder(kaldiDirectory + "/link_kaldi_tools.sh",
+                    kaldiDirectory).inheritIO().start().waitFor();
             cleanup.start();
         } catch (Exception e) {
             e.printStackTrace();
@@ -107,22 +114,22 @@ public class HomeController {
     @RequestMapping(value = "/getUserInfo", method = RequestMethod.GET)
     @ResponseBody
     public Map<String, List<String>> getUserInfo() {
-        String id = RequestContextHolder.currentRequestAttributes().getSessionId();
         String[] sources = loader.getSources();
         Map<String, List<String>> enrolledRecords = new HashMap<>();
         for(String source : sources){
-            String userDir = String.format("%s%s/data/%s", speakersDirectory, source, id);
+            String userDir = getUserDirectory(source);
             File directory = new File(userDir);
             if (!directory.exists()) {
                 directory.mkdir();
                 new File(userDir + "/models").mkdir();
             }
-            File[] records = directory.listFiles();
+            File[] files = directory.listFiles();
             List<String> recordNames = new ArrayList<>();
-            for(File record : records){
-                String name = record.getName();
-                String extensionRegex = "[.][^.]+$";
-                recordNames.add(name.replaceFirst(extensionRegex, ""));
+            for(File file : files){
+                String name = file.getName();
+                if(name.startsWith("ali.")){
+                    recordNames.add(name.substring(4));
+                }
             }
             enrolledRecords.put(source, recordNames);
         }
@@ -133,38 +140,35 @@ public class HomeController {
     @ResponseBody
     public String process(@RequestParam String source, @RequestParam("data") MultipartFile multipartFile)
             throws IOException, InterruptedException {
-        String userId = RequestContextHolder.currentRequestAttributes().getSessionId();
+        String userId = getUserId();
         String name = multipartFile.getOriginalFilename();
-        String userDataDir = String.format("%s%s/data/%s/", speakersDirectory, source, userId);
+        String userDataDir = getUserDirectory(source) + "/";
         String recordPath = userDataDir + name;
         File uploadedFile = new File(recordPath);
         multipartFile.transferTo(uploadedFile);
-        recordPath = RecordPreprocessor.preprocess(recordPath);
-        File recordFile = new File(recordPath);
-
-        String extensionRegex = "[.][^.]+$";
-        String recordId = name.replaceFirst(extensionRegex, "");
+        if(RecordPreprocessor.preprocessForKaldi(recordPath) == null){
+            return "failed";
+        }
+        String recordId = name.replaceFirst(RecordPreprocessor.ExtensionRegex, "");
 
         File text = new File(userDataDir + "text");
-        String sourceLang = source.split("_")[0];
-        String textContent = String.format("%s %s\n", recordId, texts.get(sourceLang).get(Integer.parseInt(recordId)));
+        String textContent = String.format("%s %s\n", recordId, texts.get(getLanguage(source)).get(Integer.parseInt(recordId)));
         try(Writer out = new BufferedWriter(new OutputStreamWriter(new FileOutputStream(text), "UTF8"))){
             out.write(textContent);
         }
         Path spk2utt = Paths.get(userDataDir + "spk2utt");
-        Files.write(spk2utt, String.format("0 %s\n", recordId).getBytes());
+        Files.write(spk2utt, String.format("%s %s\n", USER_DICTOR_NUMBER, recordId).getBytes());
 
         Path utt2spk = Paths.get(userDataDir + "utt2spk");
-        Files.write(utt2spk, String.format("%s 0\n", recordId).getBytes());
+        Files.write(utt2spk, String.format("%s %s\n", recordId, USER_DICTOR_NUMBER).getBytes());
 
         Path wavscp = Paths.get(userDataDir + "wav.scp");
-        String relativeName = String.format("data/%s/%s", userId, recordFile.getName());
+        String relativeName = String.format("data/%s/%s", userId, name);
         Files.write(wavscp, String.format("%s %s", recordId, relativeName).getBytes());
 
-        Process process = new ProcessBuilder("/kaldi/process.sh",
-                "/kaldi", speakersDirectory + source, userId, recordId).inheritIO().start();
+        Process process = new ProcessBuilder(kaldiDirectory + "/process.sh",
+                kaldiDirectory, speakersDirectory + source, userId, recordId).inheritIO().start();
         int result = process.waitFor();
-        uploadedFile.delete();
         return result == 0 ? "accepted" : "failed";
     }
 
@@ -175,10 +179,9 @@ public class HomeController {
         File testFolder = new File(foldername);
         File[] records = testFolder.listFiles();
         String[][] result = new String[records.length][2];
-        String extensionRegex = "[.][^.]+$";
         for (int i = 0; i < records.length; i++) {
             String filename = records[i].getName();
-            String speaker = filename.replaceFirst(extensionRegex, "");
+            String speaker = filename.replaceFirst(RecordPreprocessor.ExtensionRegex, "");
             result[i] = new String[2];
             result[i][0] = filename;
             result[i][1] = speakers.get(source).get(Integer.valueOf(speaker));
@@ -191,9 +194,10 @@ public class HomeController {
     public String recognize(@RequestParam String source, @RequestParam String path) {
         String result = "not found";
         try {
-            output.writeBytes(String.format("%s %s%s\n", source, speakersDirectory, path));
-            int speaker = Integer.valueOf(input.readLine());
-            result = speakers.get(source).get(speaker);
+            String command = getRecognizeCommand(source, speakersDirectory + path);
+            output.writeBytes(command);
+            String answer = input.readLine();
+            result = getSpeakerName(answer, source);
         } catch (Exception e) {
             e.printStackTrace();
         }
@@ -205,20 +209,57 @@ public class HomeController {
     public String recognizeRecorded(@RequestParam("source") String source,
                                     @RequestParam("data") MultipartFile multipartFile) {
         String time = new SimpleDateFormat("HH-mm-ss.SSS").format(new Date());
-        String path = String.format("%s%d_%s.wav", speakersDirectory, Thread.currentThread().getId(), time);
-        String result = null;
+        String name = multipartFile.getOriginalFilename();
+        String extension = name.substring(name.lastIndexOf('.'));
+        String path = String.format("%s%d_%s%s", speakersDirectory, Thread.currentThread().getId(), time, extension);
+        String result = "not found";
         try {
             File uploadedFile = new File(path);
             multipartFile.transferTo(uploadedFile);
             path = RecordPreprocessor.preprocess(path);
+            if(path == null){
+                return result;
+            }
             uploadedFile.delete();
-            output.writeBytes(String.format("%s %s\n", source, path));
-            int speaker = Integer.valueOf(input.readLine());
-            result = speakers.get(source).get(speaker);
+            String command = getRecognizeCommand(source, path);
+            output.writeBytes(command);
+            String answer = input.readLine();
+            result = getSpeakerName(answer, source);
             new File(path).delete();
         } catch (Exception e) {
             e.printStackTrace();
         }
         return result;
+    }
+
+    @RequestMapping(value = "/add", method = RequestMethod.GET)
+    @ResponseBody
+    public String add(@RequestParam String source) throws IOException, InterruptedException {
+        int numberOfRecords = texts.get(getLanguage(source)).size();
+        String userDir = getUserDirectory(source);
+        output.writeBytes(String.format("%s %s/ %s\n", source, userDir, numberOfRecords));
+        return input.readLine();
+    }
+
+    private String getUserId(){
+        return RequestContextHolder.currentRequestAttributes().getSessionId();
+    }
+
+    private String getLanguage(String source){
+        return source.split("_")[0];
+    }
+
+    private String getUserDirectory(String source){
+        return String.format("%s%s/data/%s", speakersDirectory, source, getUserId());
+    }
+
+    private String getSpeakerName(String answer, String source){
+        return answer.equals(USER_DICTOR_NUMBER) ?
+                "you" :
+                speakers.get(source).get(Integer.valueOf(answer));
+    }
+
+    private String getRecognizeCommand(String source, String path){
+        return String.format("%s %s%s/data/%s/ %s\n", source, speakersDirectory, source, getUserId(), path);
     }
 }
